Renamed the main.cpp files inside the test folder. Merged all 4 makefiles into 1 single make file
